+++ packages/toolbars-new/web/src/Toolbar/RicosToolbar.tsx
@@ -59,7 +59,7 @@ interface RicosToolbarProps {
   defaultLineSpacing?: defaultLineSpacingType;
   getEditorContainer: () => Element;
   cssOverride?: RicosCssOverride;
-  configButtonsMapper?: any;
+  configButtonsOverrides?: any;
   disabled?: boolean;
 }
 
@@ -111,7 +111,7 @@ class RicosToolbar extends Component<RicosToolbarProps> {
       defaultLineSpacing,
       getEditorContainer,
       cssOverride,
-      configButtonsMapper,
+      configButtonsOverrides,
       disabled,
     } = this.props;
     const updatedButtons = createButtonsList({
@@ -124,7 +124,7 @@ class RicosToolbar extends Component<RicosToolbarProps> {
       defaultLineSpacing,
       experiments,
       theme,
-      configButtonsMapper,
+      configButtonsOverrides,
     });
     const buttonsWithoutUnwantedSeparators =
       updatedButtons.length > 0 && this.cleanUnwantedSeparators(updatedButtons);